backend/requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
python-dotenv
pydantic-settings
pymupdf
pytesseract
Pillow
opencv-python-headless
spacy
dateparser
tqdm
loguru
aiofiles
jinja2
python-jose[cryptography]
passlib[bcrypt]
openai
